fix sql init
